bot/requirements.txt
discord-py-interactions==5.14.0
python-dotenv==1.0.1